+++ integration-test/package.json
@@ -2,7 +2,6 @@
   "name": "integration-test",
   "version": "1.0.0",
   "description": "",
-  "main": "login_ec.js",
   "scripts": {
     "start": "sh run_integration.sh",
     "test": "echo \"Error: no test specified\" && exit 1"
